init Renderer api-example app
